+++ src/index.d.ts
@@ -1,0 +1,7 @@
+declare module "*.png";
+declare module "*.gif";
+declare module "*.jpg";
+declare module "*.jpeg";
+declare module "*.css";
+declare module "*.css?raw";
+declare module "*.svg";